Update StartingUserInterface's CreateCommand method
 Update CreateCommand's Conditions
When checking the parts.Length it should check that parts.Length is not 3.
If parts.Length is not 3 write "Command not valid, Create requires a name and type of gradebook." to the Console.
 return a new GradeBook based on the provided type
If the value of parts[2] is "standard" return a newly instantiated StandardGradeBook using the name variable.
If the value of parts[2] is "ranked" return a newly instantiated RankedGradeBook using the name variable.
If the value of parts[2] doesn't match the above write the value of parts[2] followed by " is not a supported type of gradebook, please try again" to the console, then escape the method.